refactor: remove additional states
