+++ src/courseware/course/new-sidebar/SidebarTriggers.jsx
@@ -1,14 +1,18 @@
 import React, { useContext } from 'react';
 import SidebarContext from './SidebarContext';
-import SidebarTrigger from './SidebarTrigger';
-import { SidebarID } from './constants';
+import { SIDEBAR_ORDER, SIDEBARS } from './sidebars';
 
 const SidebarTriggers = () => {
   const { toggleSidebar } = useContext(SidebarContext);
 
   return (
-    <div className="d-flex ml-auto" key={SidebarID}>
-      <SidebarTrigger onClick={() => toggleSidebar(SidebarID)} key={SidebarID} />
+    <div className="d-flex ml-auto">
+      {SIDEBAR_ORDER.map((sidebarId) => {
+        const { Trigger } = SIDEBARS[sidebarId];
+        return (
+          <Trigger onClick={() => toggleSidebar(sidebarId)} key={sidebarId} />
+        );
+      })}
     </div>
   );
 };
