--- src/courseware/course/new-sidebar/constants.js
@@ -1,5 +1,0 @@
-export const SidebarID = 'NEWSIDEBAR';
-
-export const Discussions = 'Discussions';
-
-export const Notifications = 'Notifications';
