+++ src/courseware/course/new-sidebar/messages.js
@@ -1,19 +1,24 @@
 import { defineMessages } from '@edx/frontend-platform/i18n';
 
 const messages = defineMessages({
-  notificationTray: {
-    id: 'notification.tray.container',
-    defaultMessage: 'Notification tray',
-    description: 'Notification tray container',
+  discussionsTitle: {
+    id: 'discussions.sidebar.title',
+    defaultMessage: 'Discussions',
+    description: 'Title text for a forum where users are able to discuss course topics',
+  },
+  discussionNotificationTray: {
+    id: 'discussions.notification.tray.container',
+    defaultMessage: 'Discussion and Notification tray',
+    description: 'Discussion and Notification tray container',
   },
   notificationTitle: {
     id: 'notification.tray.title',
     defaultMessage: 'Notifications',
     description: 'Title text displayed for the notification tray',
   },
-  closeNotificationTrigger: {
-    id: 'notification.close.button',
-    defaultMessage: 'Close notification tray',
+  closeTrigger: {
+    id: 'tray.close.button',
+    defaultMessage: 'Close tray',
     description: 'Button for the learner to close the sidebar',
   },
   openSidebarTrigger: {
