+++ src/courseware/course/new-sidebar/sidebars/discussions-notifications/index.js
@@ -1,0 +1,2 @@
+export { default as Sidebar } from './DiscussionsNotificationsSidebar';
+export { default as Trigger, ID } from './DiscussionsNotificationsTrigger';
