+++ src/courseware/course/new-sidebar/sidebars/index.js
@@ -1,0 +1,13 @@
+import * as discussionsNotifications from './discussions-notifications';
+
+export const SIDEBARS = {
+  [discussionsNotifications.ID]: {
+    ID: discussionsNotifications.ID,
+    Sidebar: discussionsNotifications.Sidebar,
+    Trigger: discussionsNotifications.Trigger,
+  },
+};
+
+export const SIDEBAR_ORDER = [
+  discussionsNotifications.ID,
+];
